+++ samples/aws/lambda-sqs/SQSLambda_3/RegularEndpoint/ResponseMessageHandler.cs
@@ -4,7 +4,6 @@ using NServiceBus;
 
 public class ResponseMessageHandler(ILogger<ResponseMessageHandler> logger) : IHandleMessages<ResponseMessage>
 {
-
     public Task Handle(ResponseMessage message, IMessageHandlerContext context)
     {
         logger.LogInformation("Handling {MessageType} in RegularEndpoint", nameof(ResponseMessage));
